+++ src/test/java/org/jabref/logic/integrity/IntegrityCheckTest.java
@@ -440,7 +440,8 @@ class IntegrityCheckTest {
                 false,
                 false,
                 keyPattern,
-                ',');
+                ',',
+                false);
     }
 
     private BibDatabaseContext withMode(BibDatabaseContext context, BibDatabaseMode mode) {